test: enhance server methods integration tests
- Rename `ReifiedMethodsIT.kt` to `MokksyServerMethodsIT.kt`.
- Add additional test cases for server methods with stub configurations.
- Apply `@JvmName` annotation for better Java interoperability.
